Renamed and reorganized API files
